+++ src/Compilador/tree/Divide.java
@@ -9,7 +9,7 @@ public class Divide extends Tree {
         right = r;
     }
 
-    public Integer eval(Environment e) {
+    public Double eval(Environment e) {
         return left.eval(e) / right.eval(e);
     }
