+++ src/Compilador/tree/Plus.java
@@ -9,9 +9,10 @@ public class Plus extends Tree {
         right = r;
     }
 
-    public Integer eval(Environment e) {
+    public Double eval(Environment e) {
         return left.eval(e) + right.eval(e);
     }
+       
     public void print() {
         System.out.print("(PLUS ");
         left.print();
@@ -19,5 +20,6 @@ public class Plus extends Tree {
         right.print();
         System.out.print(")");
     }
+
 }
